Refactor(connection): now all of the typeorm... ...config is inside the connection file.
